Add DotVVM (#352)
DotVVM is an open source Knockout JS based framework that lets you build modern web apps and SPAs using C# and HTML without writing any JavaScript code.
It is been developed for more than 2 years by now, in July 2016 there was a first stable release (1.0) and currently we have the version 1.1 RC with .NET Core support.

This framework is an interesting alternative to using ASP.NET MVC in combination with Angular, React or any other client-side framework. It is also interesting for client app developers because it uses the MVVM pattern which they are usually familiar with. It doesn't require to learn 20 different JS libraries, tools or frameworks to create a web app, the framework handles all the common tasks like client-server communication, validation, localization and SPAs and includes built-in controls like GridView, FileUpload etc.

The framework is open source and there is also a free Visual Studio extension which adds IntelliSense and project templates for DotVVM apps.
(My company also develops commercial extensions and we also have a paid version of the VS extension with more features.)

I am the founder of the DotVVM project. I have spoken about DotVVM on several conferences in Europe and we got a really great feedback, many people and companies are already using it, so I think it may be the right time for DotVVM to be listed here.